refactor(analysis): split content complexity nesting tests
