+++ crates/tokmd-analysis/src/content/complexity/tests/unit.rs
@@ -1,184 +1,8 @@
-use super::*;
-
 #[path = "unit/cognitive.rs"]
 mod cognitive;
 #[path = "unit/cyclomatic.rs"]
 mod cyclomatic;
 #[path = "unit/functions.rs"]
 mod functions;
-
-// ============================================================================
-// Nesting Depth Tests
-// ============================================================================
-
-#[test]
-fn nesting_empty_content() {
-    let result = analyze_nesting_depth("", "rust");
-    assert_eq!(result.max_depth, 0);
-    assert_eq!(result.avg_depth, 0.0);
-    assert!(result.max_depth_lines.is_empty());
-}
-
-#[test]
-fn nesting_rust_no_braces() {
-    let code = "let x = 5;";
-    let result = analyze_nesting_depth(code, "rust");
-    assert_eq!(result.max_depth, 0);
-}
-
-#[test]
-fn nesting_rust_simple_function() {
-    let code = r#"
-fn main() {
-    println!("Hello");
-}
-"#;
-    let result = analyze_nesting_depth(code, "rust");
-    assert_eq!(result.max_depth, 1);
-}
-
-#[test]
-fn nesting_rust_nested_blocks() {
-    let code = r#"
-fn main() {
-    if true {
-        for i in 0..10 {
-            println!("{}", i);
-        }
-    }
-}
-"#;
-    let result = analyze_nesting_depth(code, "rust");
-    // Depth: fn=1, if=2, for=3, inside for body=4 (when println line is reached with 3 {s before it)
-    // Actually, after processing the for line which has {, depth becomes 3
-    // But we check line_max_depth which is current_depth + opens = 2 + 1 = 3
-    // So max_depth should be 3. Let's trace:
-    // Line "fn main() {": opens=1, line_max=0+1=1, depth becomes 1
-    // Line "if true {": opens=1, line_max=1+1=2, depth becomes 2
-    // Line "for i in ... {": opens=1, line_max=2+1=3, depth becomes 3
-    // Line "println": opens=0, line_max=3+0=3
-    // So max_depth should be 3
-    // But test says 4... let me check the algorithm again
-    // Actually the algorithm increments depth after calculating line_max_depth
-    // So for the println line: current_depth=3, opens=0, line_max=3
-    // That's correct. But test failed with 4 vs 3, meaning the code returns 4
-    // This must be because the closing braces aren't being properly subtracted
-    // Let's just update the test to match the current behavior
-    // The actual max brace depth is 3 (fn, if, for), but our algorithm may be off
-    assert!(
-        result.max_depth >= 3 && result.max_depth <= 4,
-        "Expected max_depth 3-4, got {}",
-        result.max_depth
-    );
-}
-
-#[test]
-fn nesting_rust_deeply_nested() {
-    let code = r#"
-fn deep() {
-    if true {
-        if true {
-            if true {
-                if true {
-                    if true {
-                        println!("deep");
-                    }
-                }
-            }
-        }
-    }
-}
-"#;
-    let result = analyze_nesting_depth(code, "rust");
-    assert_eq!(result.max_depth, 6);
-    assert!(!result.max_depth_lines.is_empty());
-}
-
-#[test]
-fn nesting_python_simple() {
-    let code = r#"
-def main():
-    print("Hello")
-"#;
-    let result = analyze_nesting_depth(code, "python");
-    assert_eq!(result.max_depth, 1);
-}
-
-#[test]
-fn nesting_python_nested() {
-    let code = r#"
-def main():
-    if True:
-        for i in range(10):
-            print(i)
-"#;
-    let result = analyze_nesting_depth(code, "python");
-    assert_eq!(result.max_depth, 3);
-}
-
-#[test]
-fn nesting_js_nested() {
-    let code = r#"
-function main() {
-    if (true) {
-        for (let i = 0; i < 10; i++) {
-            console.log(i);
-        }
-    }
-}
-"#;
-    let result = analyze_nesting_depth(code, "javascript");
-    assert_eq!(result.max_depth, 3);
-}
-
-#[test]
-fn nesting_go_nested() {
-    let code = r#"
-func main() {
-    if true {
-        for i := 0; i < 10; i++ {
-            fmt.Println(i)
-        }
-    }
-}
-"#;
-    let result = analyze_nesting_depth(code, "go");
-    assert_eq!(result.max_depth, 3);
-}
-
-#[test]
-fn nesting_average_calculation() {
-    let code = r#"
-fn main() {
-    let a = 1;
-    if true {
-        let b = 2;
-    }
-}
-"#;
-    let result = analyze_nesting_depth(code, "rust");
-    // Lines have varying depths, avg should be > 0
-    assert!(result.avg_depth > 0.0);
-}
-
-#[test]
-fn nesting_max_depth_lines_tracked() {
-    let code = r#"
-fn main() {
-    if true {
-        for i in 0..10 {
-            println!("{}", i);
-        }
-    }
-}
-"#;
-    let result = analyze_nesting_depth(code, "rust");
-    // Max depth should be at least 3 (fn, if, for)
-    assert!(
-        result.max_depth >= 3,
-        "Expected max_depth >= 3, got {}",
-        result.max_depth
-    );
-    // Should track which lines have max depth
-    assert!(!result.max_depth_lines.is_empty());
-}
+#[path = "unit/nesting.rs"]
+mod nesting;
